players from store working as before. all functins now in rounds components
