+++ src/containers/Rounds.js
@@ -2,7 +2,7 @@
 import { connect } from "react-redux";
 
 //===========Import in the component to wrap
-import Round from "../components/show/Round";
+import Rounds from "../components/show/Rounds";
 
 // mapStateToProps is passed in the current state returns object, which gets passed in as props to the connected component
 const mapStateToProps = state => {
@@ -11,5 +11,5 @@ const mapStateToProps = state => {
     };
 };
 
-// connect up mapStateToProps with the Round component
-export default connect(mapStateToProps)(Round);
+// connect up mapStateToProps with the Rounds component
+export default connect(mapStateToProps)(Rounds);
